Extract GraphQL API Logic (#57)
* Add name for long unique constraint

* Extract TagToObjectRelation Query Building

TagToObjectRelations are going to be commonly messed with in other queries and mutations. Here we are extracting out the common structure of its mutations. In the process, we also fixed a bug that occurred when deleting them as part of a Person upsert by using deleteMany instead of delete.

* Extract Person Upsert Logic Fully

* Cleanup some unused imports

* Move Person Query to API file

* Remove log and add TODO

* Extract Location API logic

* Don't compute newly(Selected|Removed)Tags

We only need to do this once. It doesn't need to be a computed variable